[Commands] Add support for rich listener objects with first-class metadata
This adds support for listener objects which, in addition to the
existing callback listeners, can optionally provide a `displayName`
ahead of time to avoid a potentially awkward humanized displayName.

Every object-based listener must implement `handleEvent()`, which plays
the role of the callback fired when the command is dispatched.

We can support other first-class metadata, for example I'm preparing a pull
request to the command palette to support displaying a command's
`description` field as a second line below the `displayName` in search
results. I even have an example of adding "tags" to a command to help
capture keywords a user might use when searching for a command, kind of
like how macOS shows "System Preferences" as a result for settings:

Here's an example of an object command listener:

```js
atom.commands.add('foo:command', document.body, {
  displayName: 'My awesome foo command',
  description: 'Here\'s some detail about my foo command. Try it out!',
  handleEvent() {
    alert('foo!');
  },
```

It should also be technically possible to attach these properties to a
function callback. If we'd like to officially support this I'll add some
tests to check:

```js
function myListener() {
  alert('foo!');
}

myListener.displayName = 'My Awesome Listener';
myListener.description = 'Does awesome stuff!';
```

I'll cross-link my command palette pull request which demonstrates the
use of this API.